code structure change and merges
Merged mlflow logging and batch prediction code to one .py file -
mlflow_batch_prediction.py. Also, created a new 02_merge_model_weights.py file to merge model weights with hf and volume instead of having it in train_llm.py, because it loads the model twice and we have to manually restart the cluster to load the model the second time. Depreciated some of the .py files as they are redundant codes and have been merged to 03_mlflow_model_batch_prediction.py